\name{poisDPmix-package}
\alias{poisDPmix-package}
\alias{poisDPmix}
\docType{package}
\title{
poisDPmix: Fit Dirichlet process mixtures of spatial Poisson process
}
\description{
Defines functions and classes to fit the models used in my disseration.
}
\details{
\tabular{ll}{
Package: \tab poisDPmix\cr
Type: \tab Package\cr
Version: \tab 0.2 alpha\cr
Date: \tab 2018-01-03\cr
License: \tab What license is it under?\cr
}

}
\author{
K.A. Flagg

Maintainer: K.A. Flagg <kenneth.flagg@msu.montana.edu>
}
\references{
Carvallo, Lopes, Polson, and Taddy (2010)

Escobar and West (1995)
}

\keyword{ package }
%\seealso{
% Optional links to other man pages, e.g. 
%}
%\examples{
%A <- matrix(rnorm(10000), 100, 100)
%B <- matrix(rnorm(10000), 100, 100)
%system.time(AB <- mat_mult(A,B))
%}
